Move services to program
Services are now provided on a per-program basis. This means that a
program depends on services, rather than vice-versa. Futhermore,
services are local to a program and its analysis, so this is a better
expression of that.
