+++ editor/language_server/src/program/services.rs
@@ -1,0 +1,7 @@
+//! Module `services` describes language services provied by a [`Program`](crate::Program) to
+//! answer server queries.
+
+mod definitions;
+mod highlight;
+mod hover;
+mod references;
